Add simplified bootstrap grid classes for migration
This might be a hot take, but I've done this to restore some semblance
of how the grids used to look, because (IMO) it's a lot easier to
iterate by preserving an existing layout with a visual aid.

I reverted some changes to move grid widths into element-specific
styles because I think those stragglers would be confusing given the
above.